fix upload file

+++ database/migrations/2022_01_10_152558_create_loan_details_table.php
@@ -17,7 +17,7 @@ class CreateLoanDetailsTable extends Migration
         Schema::create('loan_details', function (Blueprint $table) {
             $table->increments("loan_det_id");
             $table->double("amount");
-            $table->integer("interest");
+            $table->double("interest");
             $table->integer("term");
             $table->text("purpose");
             $table->unsignedBigInteger("personal_id");
